Modularize game_object
